Fix meeting link generator for production
Okay, looks like I was wrong again.
This new arrangement looks for a REACT_APP_VERCEL_URL environment
variable. If it exists, that's the base URL that's used. If that is
undefined, then the app looks for a custom REACT_APP_BASE_URL variable.
If that's undefined, we just use localhost:3000 because that's a
sensible default.